+++ tf2_sensor_msgs/include/tf2_sensor_msgs/tf2_sensor_msgs.h
@@ -30,70 +30,8 @@
 #ifndef TF2_SENSOR_MSGS_H
 #define TF2_SENSOR_MSGS_H
 
-#include <tf2/convert.h>
-#include <tf2/time.h>
-#include <sensor_msgs/msg/point_cloud2.hpp>
-#include <sensor_msgs/point_cloud2_iterator.hpp>
-#include <Eigen/Eigen>
-#include <Eigen/Geometry>
-#include <tf2_ros/buffer_interface.h>
+#warning This header is obsolete, please include tf2_sensor_msgs/tf2_sensor_msgs.hpp instead
 
-namespace tf2
-{
-
-/********************/
-/** PointCloud2    **/
-/********************/
-
-// method to extract timestamp from object
-template <>
-inline
-tf2::TimePoint getTimestamp(const sensor_msgs::msg::PointCloud2& p) {return tf2_ros::fromMsg(p.header.stamp);}
-
-// method to extract frame id from object
-template <>
-inline
-std::string getFrameId(const sensor_msgs::msg::PointCloud2 &p) {return p.header.frame_id;}
-
-// this method needs to be implemented by client library developers
-template <>
-inline
-void doTransform(const sensor_msgs::msg::PointCloud2 &p_in, sensor_msgs::msg::PointCloud2 &p_out, const geometry_msgs::msg::TransformStamped& t_in)
-{
-  p_out = p_in;
-  p_out.header = t_in.header;
-  Eigen::Transform<float,3,Eigen::Affine> t = Eigen::Translation3f(t_in.transform.translation.x, t_in.transform.translation.y,
-                                                                   t_in.transform.translation.z) * Eigen::Quaternion<float>(
-                                                                     t_in.transform.rotation.w, t_in.transform.rotation.x,
-                                                                     t_in.transform.rotation.y, t_in.transform.rotation.z);
-
-  sensor_msgs::PointCloud2ConstIterator<float> x_in(p_in, std::string("x"));
-  sensor_msgs::PointCloud2ConstIterator<float> y_in(p_in, std::string("y"));
-  sensor_msgs::PointCloud2ConstIterator<float> z_in(p_in, std::string("z"));
-
-  sensor_msgs::PointCloud2Iterator<float> x_out(p_out, std::string("x"));
-  sensor_msgs::PointCloud2Iterator<float> y_out(p_out, std::string("y"));
-  sensor_msgs::PointCloud2Iterator<float> z_out(p_out, std::string("z"));
-
-  Eigen::Vector3f point;
-  for(; x_in != x_in.end(); ++x_in, ++y_in, ++z_in, ++x_out, ++y_out, ++z_out) {
-    point = t * Eigen::Vector3f(*x_in, *y_in, *z_in);
-    *x_out = point.x();
-    *y_out = point.y();
-    *z_out = point.z();
-  }
-}
-inline
-sensor_msgs::msg::PointCloud2 toMsg(const sensor_msgs::msg::PointCloud2 &in)
-{
-  return in;
-}
-inline
-void fromMsg(const sensor_msgs::msg::PointCloud2 &msg, sensor_msgs::msg::PointCloud2 &out)
-{
-  out = msg;
-}
-
-} // namespace
+#include <tf2_sensor_msgs/tf2_sensor_msgs.hpp>
 
 #endif // TF2_SENSOR_MSGS_H
