+++ tf2_sensor_msgs/test/test_tf2_sensor_msgs.cpp
@@ -28,7 +28,7 @@
  */
 
 
-#include <tf2_sensor_msgs/tf2_sensor_msgs.h>
+#include <tf2_sensor_msgs/tf2_sensor_msgs.hpp>
 #include <geometry_msgs/PoseStamped.h>
 #include <tf2_ros/transform_listener.h>
 #include <ros/ros.h>
